requirements.txt
blinker==1.9.0
Bootstrap-Flask==2.4.1
click==8.1.7
dnspython==2.7.0
dominate==2.9.1
email_validator==2.2.0
Flask==3.1.0
Flask-WTF==1.2.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
visitor==0.1.3
Werkzeug==3.1.3
WTForms==3.2.1
